requirements.txt
discord~=1.0.1
asyncpg~=0.21.0
python-dotenv~=0.15.0